% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cli.R
\name{api_cli}
\alias{api_cli}
\title{api_cli}
\value{
a \code{api_cli} class (R6 class)
}
\description{
R6 class for handling NAMC cli driven api calls
}
\examples{

auth_config = list(...) # api_cli public or private variables
auth = api_cli$new(api,args)

Starts an api cli instance

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-start}{\code{api_cli$start()}}
\item \href{#method-clone}{\code{api_cli$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_cli$start(
  api_endpoint = NA,
  args = list(),
  include = c(),
  exclude = c(),
  filter = list(),
  sort = c(),
  limit = NA,
  api = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_endpoint}}{String name of endpoint}

\item{\code{args}}{List of named arguments to pass to the endpoint}

\item{\code{include}}{List of fields to include}

\item{\code{exclude}}{List of fields to exclude}

\item{\code{filter}}{List of fieldname / filter combinations}

\item{\code{sort}}{Array of fields to sort by}

\item{\code{limit}}{Number of results to limit the return by}

\item{\code{api}}{NAMC api instance}

\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing is returned
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_cli$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
