% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_connection}
\alias{diagnose_connection}
\title{Diagnose connection issues}
\usage{
diagnose_connection()
}
\value{
none No return
}
\description{
Diagnose connection issues
}
\examples{

NAMCr::diagnose_connection()

}
