% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize package}
\usage{
.onLoad(libnam, pkgname)
}
\arguments{
\item{libnam}{Name of the library the package is installed in.}

\item{pkgname}{The name of the package (NAMCr)}
}
\value{
none No return
}
\description{
Initialize package
}
