% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_info.R
\name{endpoint_fields}
\alias{endpoint_fields}
\title{List all endpoint fields}
\usage{
endpoint_fields(api_endpoint)
}
\arguments{
\item{api_endpoint}{The name of an API endpoint.}
}
\value{
A character vector of field names.
}
\description{
\code{endpoint_fields} retrieves the fields associated with an API endpoint from the schema#'
}
\examples{

endpoint_fields( "sites" )

}
