% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_info.R
\name{endpoint_parameters}
\alias{endpoint_parameters}
\title{List all endpoint parameters}
\usage{
endpoint_parameters(api_endpoint, no_paging = FALSE)
}
\arguments{
\item{api_endpoint}{The name of an API endpoint.}

\item{no_paging}{Logical that omits the paging arguments by default}
}
\value{
A character vector of parameters.
}
\description{
\code{endpoint_parameters} returns a vector of parameters associated with an API endpoint from the schema
}
\examples{

endpoint_parameters( "sites" )

}
