% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{execute_api_fn}
\alias{execute_api_fn}
\title{Execute public API class method}
\usage{
execute_api_fn(fn_name)
}
\arguments{
\item{fn_name}{Name of namc_api public method}
}
\value{
variable Depends on function executed. See public methods of namc_api
}
\description{
Execute public API class method
}
