% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{format_arguments}
\alias{format_arguments}
\title{format user passed arguments for graphql queries}
\usage{
format_arguments(api_endpoint, endpoint_args, api)
}
\arguments{
\item{api_endpoint}{string A NAMC API endpoint name. Available endpoints can
be found via the \code{\link[=endpoints]{endpoints()}} function.}

\item{endpoint_args}{a list of named arguments to pass to the API endpoint. Can also
be passed individually in  ...}

\item{api}{instance of a namc_api class}
}
\value{
graphql formatted query string
}
\description{
format user passed arguments for graphql queries
}
