% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{formatter}
\alias{formatter}
\title{formatter}
\value{
a \code{formatter} class (R6 class)
}
\description{
R6 class for formatting strings for console or markup output
}
\examples{

f = formatter$new()
formatted_text = f$wrap_text("text","h1","example ")

Initialize object


## ------------------------------------------------
## Method `formatter$wrap_text`
## ------------------------------------------------


txt = formatter::wrap_text("My Title","h1","")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{formatter$new()}}
\item \href{#method-wrap_text}{\code{formatter$wrap_text()}}
\item \href{#method-clone}{\code{formatter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formatter$new(level = "INFO")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The text message type to initialize messaging to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none Nothing is returned

format a string for console or markup output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wrap_text"></a>}}
\if{latex}{\out{\hypertarget{method-wrap_text}{}}}
\subsection{Method \code{wrap_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formatter$wrap_text(txt, style, appendTo = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{txt}}{string to wrap in a style}

\item{\code{style}}{string name of a styles key}

\item{\code{appendTo}}{sting to appended the styled text to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
txt = formatter::wrap_text("My Title","h1","")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formatter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
