% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_api_var}
\alias{get_api_var}
\title{Get API class variable}
\usage{
get_api_var(var_name)
}
\arguments{
\item{var_name}{Name of namc_api variable name}
}
\value{
variable Depends on variable requested. See variables of namc_api
}
\description{
Get API class variable
}
