% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_endpoint_args}
\alias{get_endpoint_args}
\title{Retrieve fields}
\usage{
get_endpoint_args(api_endpoint)
}
\arguments{
\item{api_endpoint}{The name of an API endpoint.}
}
\value{
vector A character vector of arguments.
}
\description{
Retrieves the fields associated with an API endpoint from the schema
}
\examples{

args = NAMCr::get_endpoint_args( api_endpoint = "sites" )

}
