% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_endpoint_fields}
\alias{get_endpoint_fields}
\title{Retrieve fields}
\usage{
get_endpoint_fields(api_endpoint)
}
\arguments{
\item{api_endpoint}{The name of an API endpoint.}
}
\value{
vector A character vector.
}
\description{
Retrieves the fields associated with an API endpoint from the schema
}
\examples{

endpoints = NAMCr::get_endpoints( api_endpoint = "sites" )

}
