% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_endpoints}
\alias{get_endpoints}
\title{Retrieve API endpoints}
\usage{
get_endpoints()
}
\value{
vector A character vector.
}
\description{
Retrieves the API endpoint names from the schema
}
\examples{

endpoints = NAMCr::get_endpoints()

}
