% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.empty}
\alias{is.empty}
\title{Checks for empty values}
\usage{
is.empty(val)
}
\arguments{
\item{val}{value or list of values to check for emptiness}
}
\value{
logical array
}
\description{
\code{is.empty} Null, NA and empty string values are all considered empty
}
\examples{

is.empty( list(NA,NULL,'','test',c(),list()) )

}
