% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{json.collapse}
\alias{json.collapse}
\title{collapse columns in a dataframe to a single JSON column}
\usage{
json.collapse(data, field, colNames)
}
\arguments{
\item{data}{dataframe containing one or more columns}

\item{field}{name to give JSON column}

\item{colNames}{a single or an array of column names to convert to JSON records}
}
\value{

}
\description{
collapse columns in a dataframe to a single JSON column
}
\examples{

# Multiple columns
data = json.expand(data, "newField", c("field1","field2"))
# OR single column
data = json.expand(data, "newField", "field1")

}
