% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{json.expand}
\alias{json.expand}
\title{expands JSON keys into dataframe columns}
\usage{
json.expand(data, fields)
}
\arguments{
\item{data}{dataframe containing one or more columns of JSON}

\item{fields}{a single or an array of column names containing JSON text}
}
\value{
dataframe with additional columns representing the JSON keys
}
\description{
expands JSON keys into dataframe columns
}
\examples{

# Multiple columns
data = json.expand(data, c("field1","field2"))
# OR single column
data = json.expand(data, "field1")

}
