% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{messages}
\alias{messages}
\title{messages}
\value{
a `messages` class (R6 class)
}
\description{
R6 class for handling package messages
}
\examples{

msg = messages$new(level = "INFO")

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{levels}}{is an array of the 4 level types.}

\item{\code{iLevel}}{is the index of the level contained in levels.
Initialize object

Allow list based function initialization (easier to integrate with config files)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{levels}}{is an array of the 4 level types.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{messages$new()}}
\item \href{#method-display}{\code{messages$display()}}
\item \href{#method-clone}{\code{messages$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{messages$new(level = "INFO")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argList}}{A 'List' of name/value pairs to be passed in as arguments.}

\item{\code{...}}{Name/Value pairs as arguments.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none Nothing is returned

Title
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-display"></a>}}
\if{latex}{\out{\hypertarget{method-display}{}}}
\subsection{Method \code{display()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{messages$display(msg, type = "INFO")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{string A message to display}

\item{\code{type}}{string The type of the message}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{messages$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
