% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.config.R
\name{pkg.config}
\alias{pkg.config}
\title{Get API parameters}
\usage{
pkg.config()
}
\value{
list A list of API and authentication parameters
}
\description{
Returns a nested list of high-level parameters required for making the API and authentication calls.
}
