% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query}
\alias{query}
\title{Query NAMC database}
\usage{
query(
  api_endpoint,
  args = list(),
  fields = c(),
  filter = list(),
  sort = c(),
  limit = NA,
  return_format = "dataframe",
  ...
)
}
\arguments{
\item{api_endpoint}{string A NAMC API endpoint name. Available endpoints can be found via the
namc_api get_endpoints method.}

\item{fields}{vector A vector of strings matching the desired api_endpoint fields to return.}

\item{filter}{list A list containing filter parameters.}

\item{sort}{list A list containing fields / directions to sort by.}

\item{limit}{numeric The number of results to desired.}

\item{return_format}{string One of 'dataframe', 'json'}
}
\value{
data.frame A dataframe containing the query results
}
\description{
Providing a desired endpoint and filtering the function retrieves data from the NAMC database. If not already
authenticated the will trigger the authentication / account creation routines.
}
\examples{

sites = NAMCr::query(
  api_endpoint = 'sites',
  fields = c('station','lat','long')),
  filter = list(
    station = c('station1','station2'),
    lat = list(
      data = c(42,46),
      type = 'between'
    )
  )
  sort = c('station','lat'),
  limit = 50
)

}
