% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{raw_query}
\alias{raw_query}
\title{Execute raw graphql query}
\usage{
raw_query(query_string)
}
\value{
data.frame A dataframe contained the query result.
}
\description{
Execute raw graphql query
}
\examples{

sites = NAMCr::raw_query(
  'query rQuery{
    sites {
      station lat long ...
    }
  }'
)

}
