% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reinitialize}
\alias{reinitialize}
\title{Reinitialize cached API state.}
\usage{
reinitialize(clear_credentials = FALSE)
}
\arguments{
\item{clear_credentials}{logical TRUE/FALSE indicating if cached API credentials should be deleted}
}
\description{
\code{reinitialize} reanalyzes the API schema and authentication information.
}
