% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{reinitiate}
\alias{reinitiate}
\title{Reinitiate the package state}
\usage{
reinitiate(clear_credentials = FALSE)
}
\value{
namc_api Package configured API object.
}
\description{
Reinitiate the package state
}
