% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{save}
\alias{save}
\title{Save data}
\usage{
save(
  api_endpoint,
  args = list(),
  api = .pkgenv$api,
  append_metadata = FALSE,
  ...
)
}
\arguments{
\item{api_endpoint}{string A NAMC API endpoint name. Available endpoints can be found via the}

\item{args}{a list of named arguments to pass to the API endpoint. Can also be passed individually in  ...
namc_api get_endpoints method.}

\item{api}{namc_api An instance of the namc_api class having a default of a pre-initialized package object}

\item{...}{named arguments to merge with data and pass to the API endpoint.}
}
\value{
data.frame An identifier of an inserted record OR a data.frame containing the modified record
}
\description{
Update NAMC database data via the API. Upsert routines are used.
}
\examples{

site = list(
  station = 'myStation',
  lat = 45.555,
  long = -123.555,
  ...
)

sites = NAMCr::save(
  api_endpoint = 'updateSite',
  args = site
)

# Alternative

sites = NAMCr::save(
  api_endpoint = 'updateSite',
  station = 'myStation',
  lat = 45.555,
  long = -123.555
)

}
\seealso{
\code{\link[=info]{info()}},\code{browseVignettes("NAMC-API-Overview")}

Other query functions: 
\code{\link{query}()},
\code{\link{raw_query}()}
}
\concept{query functions}
