% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{site_is_online}
\alias{site_is_online}
\title{API connectivity check}
\usage{
site_is_online(site_url)
}
\arguments{
\item{site_url}{is a url to check if it is accessible}
}
\value{
logical
}
\description{
API connectivity check
}
