% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_schema.R
\name{api_schema}
\alias{api_schema}
\title{api_schema}
\value{
a `namc_oauth2` class (R6 class)
}
\description{
R6 class for handling the NAMC graphql schema
}
\examples{

types = ( namc_api$new(argList=...) )$get_api_types()
schema = api_schema$new(types = types)


## ------------------------------------------------
## Method `api_schema$configure`
## ------------------------------------------------

Parse introspected schema


## ------------------------------------------------
## Method `api_schema$parse_endpoints`
## ------------------------------------------------


types = ( namc_api$new(argList=...) )$get_api_types()
schema = api_schema$new(types = types)
schema$parse_schema()

Discover info on API endpoint


## ------------------------------------------------
## Method `api_schema$parse_endpoint`
## ------------------------------------------------


types = ( namc_api$new(argList=...) )$get_api_types()
schema = api_schema$new(types = types)
schema$parse_endpoint("siteInfo","Query")

Title


## ------------------------------------------------
## Method `api_schema$is_arg_numeric`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$get_special_type_from_endpoint`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$has_edge`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$get_edge_name`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$get_endpoints`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$get_endpoint`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$get_endpoint_fields`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$get_endpoint_args`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$get_argument`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$is_argument_required`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `api_schema$is_argument`
## ------------------------------------------------

Title

}
\section{Super class}{
\code{\link[NAMCr:base_class]{NAMCr::base_class}} -> \code{api_schema}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{raw_types}}{raw list of endpoints and fields for the API
Title}

\item{\code{tpl_pagination_first}}{is the numeric first record to return}

\item{\code{tpl_pagination_offset}}{is the numeric offset of the return}

\item{\code{tpl_pagination_offset}}{is the numeric offset of the return}

\item{\code{tpl_page_fieldname}}{is the name of the field holding page data}

\item{\code{required_kind}}{is string value of required argument kind
Parse the API types}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{raw_types}}{raw list of endpoints and fields for the API
Title}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-configure}{\code{api_schema$configure()}}
\item \href{#method-parse_endpoints}{\code{api_schema$parse_endpoints()}}
\item \href{#method-parse_endpoint}{\code{api_schema$parse_endpoint()}}
\item \href{#method-is_arg_numeric}{\code{api_schema$is_arg_numeric()}}
\item \href{#method-get_special_type_from_endpoint}{\code{api_schema$get_special_type_from_endpoint()}}
\item \href{#method-has_edge}{\code{api_schema$has_edge()}}
\item \href{#method-get_edge_name}{\code{api_schema$get_edge_name()}}
\item \href{#method-get_endpoints}{\code{api_schema$get_endpoints()}}
\item \href{#method-get_endpoint}{\code{api_schema$get_endpoint()}}
\item \href{#method-get_endpoint_fields}{\code{api_schema$get_endpoint_fields()}}
\item \href{#method-get_endpoint_args}{\code{api_schema$get_endpoint_args()}}
\item \href{#method-get_argument}{\code{api_schema$get_argument()}}
\item \href{#method-is_argument_required}{\code{api_schema$is_argument_required()}}
\item \href{#method-is_argument}{\code{api_schema$is_argument()}}
\item \href{#method-is_endpoint}{\code{api_schema$is_endpoint()}}
\item \href{#method-clone}{\code{api_schema$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="NAMCr" data-topic="base_class" data-id="dump">}\href{../../NAMCr/html/base_class.html#method-dump}{\code{NAMCr::base_class$dump()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="NAMCr" data-topic="base_class" data-id="get_var">}\href{../../NAMCr/html/base_class.html#method-get_var}{\code{NAMCr::base_class$get_var()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="NAMCr" data-topic="base_class" data-id="initialize">}\href{../../NAMCr/html/base_class.html#method-initialize}{\code{NAMCr::base_class$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="NAMCr" data-topic="base_class" data-id="set_var">}\href{../../NAMCr/html/base_class.html#method-set_var}{\code{NAMCr::base_class$set_var()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure"></a>}}
\if{latex}{\out{\hypertarget{method-configure}{}}}
\subsection{Method \code{configure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$configure()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arg_name}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Parse introspected schema

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse_endpoints"></a>}}
\if{latex}{\out{\hypertarget{method-parse_endpoints}{}}}
\subsection{Method \code{parse_endpoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$parse_endpoints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
api_schema For method chaining
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
types = ( namc_api$new(argList=...) )$get_api_types()
schema = api_schema$new(types = types)
schema$parse_schema()

Discover info on API endpoint

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-parse_endpoint}{}}}
\subsection{Method \code{parse_endpoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$parse_endpoint(endpoint, special_type = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{Name of API endpoint to discover}

\item{\code{special_type}}{Name of the type of the API endpoint}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list Info structure of API details
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
types = ( namc_api$new(argList=...) )$get_api_types()
schema = api_schema$new(types = types)
schema$parse_endpoint("siteInfo","Query")

Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_arg_numeric"></a>}}
\if{latex}{\out{\hypertarget{method-is_arg_numeric}{}}}
\subsection{Method \code{is_arg_numeric()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$is_arg_numeric(endpoint, argname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argname}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_special_type_from_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-get_special_type_from_endpoint}{}}}
\subsection{Method \code{get_special_type_from_endpoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$get_special_type_from_endpoint(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_edge"></a>}}
\if{latex}{\out{\hypertarget{method-has_edge}{}}}
\subsection{Method \code{has_edge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$has_edge(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_edge_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_edge_name}{}}}
\subsection{Method \code{get_edge_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$get_edge_name(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_endpoints"></a>}}
\if{latex}{\out{\hypertarget{method-get_endpoints}{}}}
\subsection{Method \code{get_endpoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$get_endpoints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-get_endpoint}{}}}
\subsection{Method \code{get_endpoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$get_endpoint(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_endpoint_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_endpoint_fields}{}}}
\subsection{Method \code{get_endpoint_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$get_endpoint_fields(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_endpoint_args"></a>}}
\if{latex}{\out{\hypertarget{method-get_endpoint_args}{}}}
\subsection{Method \code{get_endpoint_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$get_endpoint_args(endpoint, no_paging = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{no_paging}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_argument"></a>}}
\if{latex}{\out{\hypertarget{method-get_argument}{}}}
\subsection{Method \code{get_argument()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$get_argument(endpoint, argname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argname}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_argument_required"></a>}}
\if{latex}{\out{\hypertarget{method-is_argument_required}{}}}
\subsection{Method \code{is_argument_required()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$is_argument_required(endpoint, argname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argname}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_argument"></a>}}
\if{latex}{\out{\hypertarget{method-is_argument}{}}}
\subsection{Method \code{is_argument()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$is_argument(endpoint, argname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argname}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-is_endpoint}{}}}
\subsection{Method \code{is_endpoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$is_endpoint(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_schema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
