% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_class.R
\name{base_class}
\alias{base_class}
\title{base_class}
\value{
a `base_class` class (R6 class)
}
\description{
R6 class for providing additional default R6 class features
}
\examples{
initialize - Allow list based function initialization (easier to integrate with config files)


## ------------------------------------------------
## Method `base_class$new`
## ------------------------------------------------

dump - for debugging internal class values (puts var named tmpVar in global environment)


## ------------------------------------------------
## Method `base_class$dump`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `base_class$get_var`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `base_class$set_var`
## ------------------------------------------------

.set_from_list - sets all public and private fields from a name,value list

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{base_class$new()}}
\item \href{#method-dump}{\code{base_class$dump()}}
\item \href{#method-get_var}{\code{base_class$get_var()}}
\item \href{#method-set_var}{\code{base_class$set_var()}}
\item \href{#method-clone}{\code{base_class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{base_class$new(argList = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dump - for debugging internal class values (puts var named tmpVar in global environment)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump"></a>}}
\if{latex}{\out{\hypertarget{method-dump}{}}}
\subsection{Method \code{dump()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{base_class$dump(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var"></a>}}
\if{latex}{\out{\hypertarget{method-get_var}{}}}
\subsection{Method \code{get_var()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{base_class$get_var(var_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_name}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_var"></a>}}
\if{latex}{\out{\hypertarget{method-set_var}{}}}
\subsection{Method \code{set_var()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{base_class$set_var(var_name, var_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_value}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{.set_from_list - sets all public and private fields from a name,value list

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{base_class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
