% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_info.R
\name{info}
\alias{info}
\title{Describe the NAMC API.}
\usage{
info(endpoint = NA, no_paging = TRUE, format = "text", to_console = TRUE)
}
\arguments{
\item{endpoint}{api endpoint name to get information about.}

\item{no_paging}{Logical that omits the paging arguments by default}

\item{format}{One of: 'text (default), 'markdown'}

\item{to_console}{Logical TRUE(default), FALSE to print output to console}
}
\value{
If \code{format} argument is console nothing is returned otherwise text is
returned in the format defined.
}
\description{
\code{info} prints text to the console describing the NAMC API
options.
}
\details{
This function presents a text block describing the available API
endpoints and their corresponding input parameters. Inputs are marked as
required where applicable. Endpoint descriptions can be subset by providing
an endpoint name as input.
NAMC API:
\subsection{auth}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{loggedIn}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{userPool}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{clientId}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{region}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{domain}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{organizations}{
\subsection{Description}{

:API endpoint that calls entity.fn_organizations
}

\subsection{Input Parameters}{
}

\subsection{searchTerm}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{entityId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{organizationName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{organizationType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{isLab}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{address1}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{address2}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{city}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{usState}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{country}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{zipCode}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{phone}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{fax}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{website}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{notes}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{siteInfo}{
\subsection{Description}{

:Detailed information about a single site. Includes the point location and upstream catchment geometries as GeoJSON.
}

\subsection{Input Parameters}{
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:Unique database generated integer that uniquely identifies each site
}

\subsection{siteName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:Unique text identifier for each site
}

\subsection{system}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The ecological system (e.g. lake, pond, reservoir) in which the site is located.
}

\subsection{ecosystem}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The ecosystem in which the site is located.
}

\subsection{location}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:GeoJSON point location of the site in \href{https://epsg.io/4326}{EPSG:4326} spatial reference.
}

\subsection{longitude}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE

description
:Longtidue of the site location in decimal degrees.
}

\subsection{latitude}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE

description
:Latitude of the site location in decimal degrees.
}

\subsection{usState}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:US State in which the site is located.
}

\subsection{waterbodyType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:If the site has a waterbody code, then the waterbody type identifies the system that
this identifier pertains to. The waterbody type might be NHDPlus 1:100,000 or
NHDPlusHR 1:24,000 etc.
}

\subsection{waterbodyCode}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The identifier of the waterbody on which this site occurs. Can be null. If a
waterbody code exists, then the waterbodyType can be used to determine which
system the code belongs to, such as NHDPlus or NHDPlusHR.
}

\subsection{waterbodyName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The string name of the waterbody on which the site occurs. Typicall this is the stream name.
}

\subsection{geographyChanged}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The latest date and time that either the site location (point) or catchment (polygon) were changed.
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The system generated date and time that the site record was created in the database.
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The system generated date and time that the site record was last changed in the database.
}

\subsection{catchment}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:GeoJSON polygon of the upstream area that drains into this site in \href{https://epsg.io/4326}{EPSG:4326} spatial reference.
}

\subsection{sampleCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The number of samples that are available for this site.
}

}

\subsection{modelInfo}{
\subsection{Description}{

:Detailed information about a single model, including the number of predictors associated with the model.
}

\subsection{Input Parameters}{
}

\subsection{modelId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{modelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{modelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{abbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{modelType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{translation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{extentDescription}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{platform}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{referenceSites}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{groupCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{minimumCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{oeMean}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{oeStdev}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{taxonomicEffort}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{isActive}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{fixedCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{units}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{predictorCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{extent}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{modelConditions}{
\subsection{Description}{

:List of all model conditions (e.g. good, fair, poor) for a particular model
}

\subsection{Input Parameters}{
}

\subsection{modelId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{modelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{conditionId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{condition}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{displayText}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{boxInfo}{
\subsection{Description}{

:Detailed information about a single box, including information about the customer and number of samples associated with the box.
}

\subsection{Input Parameters}{
}

\subsection{boxId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{boxId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{customerId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{customerName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{customerAbbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{submitterId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{submittedBy}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{boxStateId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{boxState}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{boxReceivedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{processingCompleteDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{projectedCompleteDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{measurements}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sorterQa}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{taxaQa}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{samples}{
\subsection{Description}{

:List of all samples in the system, including high level information about the associated box and customer.
}

\subsection{Input Parameters}{
}

\subsection{sampleIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{boxIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{projectIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{entityIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{siteIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{polygon}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{pointDistance}{

data type
:QueryDistance

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{boxId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{customerName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{customerAbbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{submittedBy}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{boxState}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{siteName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{usState}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{siteLocation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{siteLongitude}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{siteLatitude}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{visitId}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{customerSiteCode}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleTime}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleMethod}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{habitat}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleLocation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleLongitude}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{sampleLatitude}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{area}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{fieldSplit}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{fieldNotes}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{labSplit}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{jarCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{qualitative}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{labNotes}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{mesh}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleDateChanged}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{qaSampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{sites}{
\subsection{Description}{

:Summary information for all sites, including location coordinates and number of samples. Use the siteInfo query to get more detailed information about a single site.
}

\subsection{Input Parameters}{
}

\subsection{usState}{

data type
:String

is numeric
:FALSE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{sampleIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{boxIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{projectIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{entityIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{siteIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{polygon}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{pointDistance}{

data type
:QueryDistance

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{siteName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{system}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{ecosystem}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{longitude}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{latitude}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{usState}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{waterbodyType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{waterbodyCode}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{waterbodyName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{hasCatchment}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{boxes}{
\subsection{Description}{

:Summary information for all boxes, including the customer name and number of samples within the box. Use the boxInfo query to get more detailed information about a single box.
}

\subsection{Input Parameters}{
}

\subsection{boxIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{entityIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{boxId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{customerId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{customerName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{submitterId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{submittedBy}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{boxState}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{boxReceivedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{processingCompleteDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{projectedCompleteDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{projects}{
\subsection{Description}{

:Summary information about all projects in the system. Use the projectInfo query to get more detailed about a single project.
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{projectId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{projectName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{isPrivate}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{contactId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{contactName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{sampleCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{modelCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{predictors}{
\subsection{Description}{

:List of all predictors in the system. Includes summary information about how many models rely on each predictor.
}

\subsection{Input Parameters}{
}

\subsection{modelId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{predictorId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The unique system generated identifier for the predictor.
}

\subsection{predictorName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The unique textual name for the predictor.
}

\subsection{abbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The unique shorthand abbreviation for the predictor.
}

\subsection{units}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The units in which the predictor values are stored.
}

\subsection{predictorType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The predictor type (e.g. metrics, atmosphere, geology, anthro).
}

\subsection{isTemporal}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE

description
:Boolean representing whether the predictor varies over time or whether
there is just a single value for the site. True indicates that the predictor
is temporal and values are stored for each sample. False indicates that there
is only one predictor value for each site.
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:Long form information about the predictor.
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:Miscellaneou structured metadata in GeoJSON format.
}

\subsection{modelCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The number of models that use the predictor
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The system generated date and time that the site record was created in the database.
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The system generated date and time that the site record was last changed in the database.
}

\subsection{calculationScript}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:Optional name of the R function that performs the calculation for this predictor.
}

\subsection{source}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:Source information about the predictors
}

}

\subsection{models}{
\subsection{Description}{

:List of all models in the system. Includes summary information about how many predictors are used by the model as well as the translation that might be required.
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{modelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{modelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{abbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{isActive}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{predictorCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

}

\subsection{sitePredictorValues}{
\subsection{Description}{

:List of predictor values at a single site. Site predictors are non-temporal. The value is common to all samples at the specific site.
}

\subsection{Input Parameters}{
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{predictorId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The unique system generated identifier for the predictor.
}

\subsection{predictorName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The unique textual name for the predictor.
}

\subsection{abbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The unique shorthand abbreviation for the predictor.
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:Long form information about the predictor.
}

\subsection{predictorType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The predictor type (e.g. metrics, atmosphere, geology, anthro).
}

\subsection{predictorValue}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The predictor value for this site. The value is always a string, even
if it represents an integer or floating point value.
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The system generated date and time that the site record was created in the database.
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The system generated date and time that the site record was last changed in the database.
}

\subsection{calculationScript}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:Optional name of the R function that performs the calculation for this predictor.
}

}

\subsection{samplePredictorValues}{
\subsection{Description}{

:List of all predictor values for a single sample. This includes both temporal and non-temporal predictors as well as their calculation status (current, missing, expired).
}

\subsection{Input Parameters}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{predictorId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{abbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{calculationScript}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{isTemporal}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{predictorMetadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{predictorValue}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{predictorValueUpdatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{status}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{modelResults}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{
}

\subsection{sampleIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{siteName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{modelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{modelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{modelVersion}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{modelResult}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{condition}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{fixCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{notes}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{taxonomy}{
\subsection{Description}{

:List of all taxa in the system. Includes information about each level of the taxonomic hierarchy above each taxa.
}

\subsection{Input Parameters}{
}

\subsection{searchTerm}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{parentTaxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{parentScientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{parentLevelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{parentLevelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{notes}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{taxonomyTree}{
\subsection{Description}{

:list all parents for a particular taxa
}

\subsection{Input Parameters}{
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{parentId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

}

\subsection{attributes}{
\subsection{Description}{

:List of all attributes in the system.
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{attributeId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{attributeName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{attributeType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{label}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{taxaAttributes}{
\subsection{Description}{

:List of all attributes that exist for a single taxa
}

\subsection{Input Parameters}{
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{attributeName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{attributeType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{label}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{attributeValue}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{translationTaxa}{
\subsection{Description}{

:List all the taxa within a specific translation (OTU)
}

\subsection{Input Parameters}{
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{translationName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{originalScientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{translationScientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{isFinal}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{translations}{
\subsection{Description}{

:List of all translations in the system.
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{translationName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{isActive}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{taxaCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{planktonSamples}{
\subsection{Description}{

:List of all plankton samples in the system.
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{diameter}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{subSampleCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{towLength}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{volume}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{aliquot}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{sizeInterval}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{towType}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{driftSamples}{
\subsection{Description}{

:List of all drift samples in the system.
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{netArea}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{netDuration}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{streamDepth}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{netDepth}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{netVelocity}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{fishSamples}{
\subsection{Description}{

:List of all fish samples in the system.
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{fishLength}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{fishMass}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{massSamples}{
\subsection{Description}{

:List of all mass samples in the system.
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{typeId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{typeAbbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{typeName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{methodId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{methodAbbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{methodName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{mass}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{fishDiet}{
\subsection{Description}{

:List all fish diet samples in the system
}

\subsection{Input Parameters}{
}

\subsection{sampleIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{sampleDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{siteName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{fishWeight}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{fishLength}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{fishTaxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{fishScientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{notes}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{organicWeight}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{inorganicWeight}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{otherWeight}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{lifeStageId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{lifeStage}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{count}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{weight}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

}

\subsection{sampleTaxaRaw}{
\subsection{Description}{

:Sample organisms summed by taxonomy only. Includes both raw counts as well as counts corrected for lab and field split. Also includes both raw and corrected big rare counts.
}

\subsection{Input Parameters}{
}

\subsection{sampleIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{boxIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{projectIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The sample to which the taxa belong
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxa identified within the laboratory.
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The name of the taxa identified with the laboratory.
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{abundance}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE

description
:Abundance for the specified taxa and lifestage
}

}

\subsection{sampleTaxaGeneralized}{
\subsection{Description}{

:Sample organisms summed by taxonomy, life stage and bug size. Includes both raw counts as well as counts corrected for lab and field split. Also includes both raw and corrected big rare counts.
}

\subsection{Input Parameters}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{lifeStageId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{lifeStage}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{lifeStageAbbreviation}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{bugSize}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{rawCount}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{labSplit}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{fieldSplit}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{correctedCount}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{rawBigRareCount}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

}

\subsection{sampleTaxaTranslation}{
\subsection{Description}{

:Sample organisms converted to a translation (OTU). Rolls up counts to those taxa present in the translation and omits any organisms that don't roll up to a taxa in the translation.
}

\subsection{Input Parameters}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The sample to which the taxa belong
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxa identified within the laboratory.
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The name of the taxa identified with the laboratory.
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{abundance}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE

description
:Abundance for the specified taxa and lifestage
}

}

\subsection{sampleTaxaTranslationRarefied}{
\subsection{Description}{

:Sample organisms converted to the translation (OTU) and then rarefied to the specified fixed count.
}

\subsection{Input Parameters}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{fixedCount}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The sample to which the taxa belong
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxa identified within the laboratory.
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The name of the taxa identified with the laboratory.
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{abundance}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE

description
:Abundance for the specified taxa and lifestage
}

}

\subsection{pointTaxaRaw}{
\subsection{Description}{

:Sample organisms within distance (meters) of a point (decimal degrees)
}

\subsection{Input Parameters}{
}

\subsection{longitude}{

data type
:Float

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{latitude}{

data type
:Float

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{distance}{

data type
:Float

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The sample to which the taxa belong
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxa identified within the laboratory.
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The name of the taxa identified with the laboratory.
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{abundance}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE

description
:Abundance for the specified taxa and lifestage
}

}

\subsection{polygonTaxaRaw}{
\subsection{Description}{

:Sample organisms within a polygon
}

\subsection{Input Parameters}{
}

\subsection{polygon}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The sample to which the taxa belong
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxa identified within the laboratory.
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The name of the taxa identified with the laboratory.
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{levelName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE

description
:The taxonomic level of the taxa identified within the laboratory.
}

\subsection{abundance}{

data type
:Float

is numeric
:TRUE

is expandable
:FALSE

description
:Abundance for the specified taxa and lifestage
}

}

\subsection{metrics}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{

none
}

\subsection{Output Fields}{
}

\subsection{metricId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{metricName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{displayText}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{groupId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{groupName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{formulaId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{formulaName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{codeFunction}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{isStandardized}{

data type
:Boolean

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{perturbDirection}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{createdDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{updatedDate}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{sampleMetrics}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{
}

\subsection{sampleIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{boxIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{projectIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{fixedCount}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{groupId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{groupName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metricId}{

data type
:Int

is numeric
:TRUE

is expandable
:FALSE
}

\subsection{metricName}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

\subsection{metricValue}{

data type
:String

is numeric
:FALSE

is expandable
:FALSE
}

}

\subsection{setSitePredictorValue}{
\subsection{Description}{

:Store a non-temporal predictor value for a specific site. See setSamplePredictorValue if you want to store a temporal predictor value associated with a particular sample.
}

\subsection{Input Parameters}{
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{predictorId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{value}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{setSamplePredictorValue}{
\subsection{Description}{

:Store a temporal predictor value for a specific sample. Set setSitePredictorValue if you want to store a temporal predictor value associated with a particular site.
}

\subsection{Input Parameters}{
}

\subsection{sampleId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{predictorId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{value}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{setSiteCatchment}{
\subsection{Description}{

:Store the upstream catchment polygon for a specific site. The catchment polygon must be a valid, non-empty GeoJSON polygon that partially or entirely intersects with the United States.
}

\subsection{Input Parameters}{
}

\subsection{siteId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{catchment}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{createTranslation}{
\subsection{Description}{

:Create a new translation. The name is mandatory and cannot be an empty string. The description is optional.
}

\subsection{Input Parameters}{
}

\subsection{translationName}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{setTranslationTaxa}{
\subsection{Description}{

:Add or update a taxa associated with a translation
}

\subsection{Input Parameters}{
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{alias}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{isFinal}{

data type
:Boolean

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{deleteTranslationTaxa}{
\subsection{Description}{

:Permanently remove a taxa from a translation
}

\subsection{Input Parameters}{
}

\subsection{translationId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{setTaxonomy}{
\subsection{Description}{

:Update information related to a specific taxa. Change its scientific name, move its level or change its parent
}

\subsection{Input Parameters}{
}

\subsection{taxonomyId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{scientificName}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{levelId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{parentId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{author}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{year}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{notes}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{createProject}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{
}

\subsection{projectName}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{isPrivate}{

data type
:Boolean

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{description}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{metadata}{

data type
:String

is numeric
:FALSE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{contactId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{addProjectSamples}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{
}

\subsection{projectId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{sampleIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{addProjectBoxes}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{
}

\subsection{projectId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{boxIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{removeProjectSamples}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{
}

\subsection{projectId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{sampleIds}{

data type
:Int

is numeric
:TRUE

is array
:TRUE

is required
:FALSE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}

\subsection{deleteProject}{
\subsection{Description}{

:
}

\subsection{Input Parameters}{
}

\subsection{projectId}{

data type
:Int

is numeric
:TRUE

is array
:FALSE

is required
:TRUE

is for paging
:FALSE
}

\subsection{Output Fields}{

none
}

}
}
\examples{

info()

}
\seealso{
\code{browseVignettes("NAMC-API-Overview")}
}
