% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query}
\alias{query}
\title{Query NAMC database.}
\usage{
query(
  api_endpoint,
  args = list(),
  include = c(),
  exclude = c(),
  filter = list(),
  sort = c(),
  limit = NA,
  api = .pkgenv$api,
  expand_metadata = TRUE,
  default_JSON_fieldName = .pkgenv$data$default_JSON_fieldName,
  ...
)
}
\arguments{
\item{api_endpoint}{string A NAMC API endpoint name. Available endpoints can
be found via the \code{\link[=endpoints]{endpoints()}} function.}

\item{args}{a list of named arguments to pass to the API endpoint. Can also
be passed individually in  ...}

\item{include}{is a vector of fields to include from the endpoint.}

\item{exclude}{is a vector of fields to exclude from the endpoint.}

\item{filter}{list A list containing filter parameters.}

\item{sort}{list A list containing fields / directions to sort by.}

\item{limit}{numeric The number of results desired.}

\item{api}{namc_api An instance of the namc_api class having a default of a
pre-initialized package object}

\item{expand_metadata}{logical if JSON metadata field is expanded into dataframe columns}

\item{default_JSON_fieldName}{"metadata" is the preset default}

\item{...}{named arguments to merge with args and pass to the API endpoint.}
}
\value{
data.frame A dataframe containing the query results.
}
\description{
\code{query} retrieves data from NAMC given a specified API endpoint
name.
}
\details{
Providing a desired endpoint and filtering the function retrieves
data from the NAMC database. If not already authenticated the will trigger
the authentication / account creation routines. By default all fields are
included. Use include and exclude arguments to limit returned fields.
}
\examples{

sites = NAMCr::query(
  api_endpoint = 'sites',
  include = c('station','lat','long')),
  filter = list(
    station = c('station1','station2'),
    lat = list(
      data = c(42,46),
      type = 'between'
    )
  )
  sort = c('station','lat'),
  limit = 50
)

# View full API docs in console
NAMCr::info()

}
\seealso{
\code{\link[=info]{info()}},\code{browseVignettes("NAMC-API-Overview")}

Other query functions: 
\code{\link{raw_query}()},
\code{\link{save}()}
}
\concept{query functions}
