% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{save}
\alias{save}
\title{Save data}
\usage{
save(api_endpoint, data, api = .pkgenv$api)
}
\arguments{
\item{api_endpoint}{string A NAMC API endpoint name. Available endpoints can be found via the
namc_api get_endpoints method.}

\item{data}{data.frame or list The data to upsert into the database.}
}
\value{
data.frame A dataframe containing any fields added to the Query
}
\description{
Update NAMC database data via the API. Upsert routines are used.
}
\examples{

site = list(
  station = 'myStation',
  lat = 45.555,
  long = -123.555,
  ...
)

sites = NAMCr::save(
  api_endpoint = 'sites',
  data = site
)

}
